# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Kudos",
        "description": "Demo Nova Kudos Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/profiles/users/companies/{company_id}/avatar": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "update profile user",
                "description": "update profile user",
                "operationId": "update profile user",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "file",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/profiles/users/companies/{company_id}/avatar"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "validation error"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profiles/users/update/mobile": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "update mobile",
                "description": "update mobile",
                "operationId": "update mobile",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09350000001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/profiles/companies/{company_id}/update/mobile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profiles/users/verify/mobile": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "verify mobile",
                "description": "verify mobile",
                "operationId": "verify mobile",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "otp_code"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09350000001"
                                    },
                                    "otp_code": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/profiles/companies/{company_id}/update/mobile"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}